backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ashatheo <ashatheo@student.21-school.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^3.6.0"
pydantic-settings = "^2.2.1"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
tortoise-orm = "^0.21.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
